Move some files to proper location
